requirements/base.txt
# Basic stuff
django>=1.8,<1.9

# Configuration
django_dotenv==1.3.0
dj-database-url==0.3.0

# Security
django-braces==1.4.0
django-secure==1.0.1

# Users
django-allauth==0.19.1
django-crispy-forms==1.4.0

# API
djangorestframework==3.1.1
markdown==2.6.1

# Astronomy
orbit==0.2
